Moving IntrusiveList to iree/hal/vulkan/util/, the last user. Only the timeline semaphore emulation uses this so I'm moving it to a place where it will be easy to delete (and attribute binary size to).
