+++ iree/base/CMakeLists.txt
@@ -262,33 +262,6 @@ flatbuffer_c_library(
 
 iree_cc_library(
   NAME
-    intrusive_list
-  HDRS
-    "intrusive_list.h"
-    "intrusive_list_ref_ptr.inc"
-    "intrusive_list_unique_ptr.inc"
-  DEPS
-    ::logging
-    ::ref_ptr
-  PUBLIC
-)
-
-iree_cc_test(
-  NAME
-    intrusive_list_test
-  SRCS
-    "intrusive_list_ref_ptr_test.cc"
-    "intrusive_list_test.cc"
-    "intrusive_list_unique_ptr_test.cc"
-  DEPS
-    ::intrusive_list
-    absl::memory
-    iree::testing::gtest
-    iree::testing::gtest_main
-)
-
-iree_cc_library(
-  NAME
     logging
   HDRS
     "logging.h"
