+++ iree/hal/vulkan/BUILD
@@ -93,13 +93,13 @@ cc_library(
         "//iree/base:arena",
         "//iree/base:core_headers",
         "//iree/base:flatcc",
-        "//iree/base:intrusive_list",
         "//iree/base:logging",
         "//iree/base:ref_ptr",
         "//iree/base:status",
         "//iree/base:synchronization",
         "//iree/base:tracing",
         "//iree/hal:api",
+        "//iree/hal/vulkan/util:intrusive_list",
         "//iree/schemas:spirv_executable_def_c_fbs",
         "@com_google_absl//absl/base:core_headers",
         "@com_google_absl//absl/container:flat_hash_map",
