+++ iree/hal/vulkan/util/BUILD
@@ -1,0 +1,56 @@
+# Copyright 2019 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+package(
+    default_visibility = ["//visibility:public"],
+    features = ["layering_check"],
+    licenses = ["notice"],  # Apache 2.0
+)
+
+iree_cmake_extra_content(
+    content = """
+if(NOT ${IREE_HAL_DRIVER_VULKAN})
+  return()
+endif()
+""",
+)
+
+
+cc_library(
+    name = "intrusive_list",
+    hdrs = [
+        "intrusive_list.h",
+        "intrusive_list_ref_ptr.inc",
+        "intrusive_list_unique_ptr.inc",
+    ],
+    deps = [
+        "//iree/base:logging",
+        "//iree/base:ref_ptr",
+    ],
+)
+
+cc_test(
+    name = "intrusive_list_test",
+    srcs = [
+        "intrusive_list_ref_ptr_test.cc",
+        "intrusive_list_test.cc",
+        "intrusive_list_unique_ptr_test.cc",
+    ],
+    deps = [
+        ":intrusive_list",
+        "//iree/testing:gtest",
+        "//iree/testing:gtest_main",
+        "@com_google_absl//absl/memory",
+    ],
+)
