+++ iree/hal/vulkan/util/CMakeLists.txt
@@ -1,0 +1,42 @@
+# Copyright 2019 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+iree_add_all_subdirs()
+
+iree_cc_library(
+  NAME
+    intrusive_list
+  HDRS
+    "intrusive_list.h"
+    "intrusive_list_ref_ptr.inc"
+    "intrusive_list_unique_ptr.inc"
+  DEPS
+    iree::base::logging
+    iree::base::ref_ptr
+  PUBLIC
+)
+
+iree_cc_test(
+  NAME
+    intrusive_list_test
+  SRCS
+    "intrusive_list_ref_ptr_test.cc"
+    "intrusive_list_test.cc"
+    "intrusive_list_unique_ptr_test.cc"
+  DEPS
+    ::intrusive_list
+    absl::memory
+    iree::testing::gtest
+    iree::testing::gtest_main
+)
