+++ iree/hal/vulkan/util/intrusive_list_unique_ptr_test.cc
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include "absl/memory/memory.h"
-#include "iree/base/intrusive_list.h"
+#include "iree/hal/vulkan/util/intrusive_list.h"
 #include "iree/testing/gtest.h"
 
 namespace iree {
